fixes after rebasing onto main
