--- elkapod_controller_gui/package.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>elkapod_controller_gui</name>
-  <version>1.0.0</version>
-  <description>Elkapod controller GUI</description>
-  <maintainer email="piotrpatek17@gmail.com">Piotr Patek</maintainer>
-  <license>MIT</license>
-
-  <depend>rclpy</depend>
-  <depend>tf_transformations</depend>
-  <exec_depend>elkapod_msgs</exec_depend>
-  <exec_depend>std_msgs</exec_depend>
-  <exec_depend>std_srvs</exec_depend>
-  <exec_depend>nav_msgs</exec_depend>
-  <exec_depend>geometry_msgs</exec_depend>
-  <exec_depend>rtabmap_msgs</exec_depend>
-  <export>
-    <build_type>ament_python</build_type>
-  </export>
-</package>
